Fixed extra window showing at startup
A second window was being shown at startup for an unknown reason. The
reason turned out to be that the window.ui file specified a GtkWindow as
the ui element, instead of a template for the SoliWindow class. Changing
the interface ui-file to match the GTK+ example application removed the
unwanted second window.