Fix SoliDocument error & crash.
The issue was that the SoliViewFrame would create its SoliView from the
template and that view would somehow create its own TextViewBuffer. In
Gedit, the automatically created buffer was of the derived type,
GeditDocument.
The key to creating a SoliDocument in the SoliView by default was to
override the create_buffer function in the TextViewBufferClass.
Returning a new SoliDocument (casted to a TextViewBuffer) from the
overriden function made sure that the buffer would be of the correct
(derived) type.